requirements.txt
google-cloud-vision
openai==0.28
requests
beautifulsoup4